Re-integrating the python-dev feature branch.
This upgrades a number of the python 3rdParty libraries, and updates 
the evaluation scripts. It introduces a new script to plot thumbnails, 
and a new diagnostics script to make use of it.
This addresses the following tickets: ASKAPSDP-2382, ASKAPSDP-2503 and 
ASKAPSDP-2543

+++ Code/Components/Analysis/evaluation/current/askap/analysis/evaluation/distributionPlots.py
@@ -543,7 +543,7 @@ def spatPosPlot(xS=None, yS=None, xR=None, yR=None, matchFlag=None, xMiss=None, 
     rc('ytick', labelsize=6)
 #    axes(setBox(232,'r'))
     axes(setBox(locationCode,'r'))
-    n, bins, patches = hist(offset, max(minPlottedOff,max(offset))/0.25)
+    n, bins, patches = hist(offset, int(max(minPlottedOff,max(offset))/0.25))
     ax=axis()
 #    ax2=[ax[0],ax[1],ax[2],max(minPlottedOff,max(offset))]
 #    axis(ax2)
